fix: S1-S8 coverage gap fixes + mutexio analyzer (round 4)
Systematic codebase review (104 findings) fixed 96 issues across
concurrency, security, lifecycle, and correctness domains:

- Critical: HTTP auth header on all client methods, piper temp-file
  use-after-remove, debug session mutex + drain lifecycle
- High: team/pair orchestrator mutexes, bus atomic counter, SSRF
  dial-time IP revalidation, predictable IDs (4 locations), git_sync
  context cancellation, STT/PTY subprocess reaping
- Medium: 39 fixes across agent, tools, LLM, security, cluster
- Low: 38 fixes including deferred S6-18..S6-25 (PTY kill, events
  trim, runtime timeout, piper relock, errcls Temporary, launchd),
  S7-3/S7-4 (skills frontmatter, TLS warning)
- Bonus: pre-existing race in ContextCompactor.Compact (found via
  verification, not original review)

New tool: tools/analyzers/mutexio/ — go/analysis analyzer enforcing
the CLAUDE.md mutex-scope rule (no I/O under lock). Run via
`make mutexio`.

8 deferred (config plumbing, design decisions, Docker API limit).

All packages build, vet, and pass -race tests.

+++ internal/repomap/graph.go
@@ -130,11 +130,13 @@ type weightedLine struct {
 
 // newWeightedLine creates a new weighted line with a unique ID.
 func newWeightedLine(from, to graph.Node, weight float64) *weightedLine {
+	// Pack two int64 node IDs into a single unique ID using integer arithmetic.
+	// This requires node IDs to stay well below 1e9 (1,000,000,000) to avoid overflow.
 	return &weightedLine{
 		from:   from,
 		to:     to,
 		weight: weight,
-		IDVal:  from.ID()*1e9 + to.ID(), // Generate unique ID
+		IDVal:  from.ID()*int64(1e9) + to.ID(), // Generate unique ID
 	}
 }
